Add exceptions to catpture error spans
An exception can be thrown during a parse at a point when the parser is
known to have failed. The exception can be caught by the combinator that
marks the start of the element that was being parsed. This the sets the
associated span of the error to the element that was being parsed.