Bind errors to a particulat parse outcome
To allow errors to be associated with particular spans and to be
associated with particular parse results.

This allows an error to be attached to a particular region of code.

This also ensures that duplicate errors will not be produced as only
those that relate to complete and successful parses will be retained.